src/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
black==22.12.0
click==8.1.3
Django==4.1.4
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.14.0
isort==5.11.3
mypy-extensions==0.4.3
pathspec==0.10.3
Pillow==9.3.0
platformdirs==2.6.0
pytz==2022.7
sqlparse==0.4.3
tomli==2.0.1
typing-extensions==4.4.0
